Create POLIN.cpp
Sneil is given NN points on a coordinate graph. From these points (X_i, Y_i)(Xi ,Yi ) he can draw horizontal and vertical lines. If he finds the number of lines that can be drawn he will get to eat Sneilium.
Since Sneil is too slow he asks you to help him with this task.

After we have drawn all the lines, we have a grid with some lines parallel to the XX axis and some lines parallel to the YY axis. The total number of lines is just the sum of the lines parallel to the XX axis and the lines parallel to the YY axis.

The only catch is to handle duplicates. For that, we can use set data structure. We club all the XX coordinates and all the YY coordinates into separate sets.

Finally, the total number of lines is just the sum of the sizes of both these sets.

Ordered sets have a complexity of O(logN)O(logN) per insertion. This leads to a time complexity of O(NlogN)O(NlogN) per testcase. We can use unordered sets to reduce this to O(N)O(N) per test case.