Create EQUALMEX.cpp
You are given an array AA containing 2\cdot N2⋅N integers. Is it possible to reorder the elements of the array in such a way so that the MEX of the first NN elements is equal to the MEX of last NN elements?

Let us denote the subarray with first NN elements as PP and that with last NN elements as QQ.

We know that, for a particular number ii to exist as the MEX of PP as well as QQ, all jj (0 \leq j < i)(0≤j<i) must have atleast one occurence in PP as well as atleast one occurence in QQ. This means that there must be atleast two occurrences of jj in the array AA.

Let ii be the smallest number such that, the number of occurence of ii in AA is less than 22.

There are 00 occurrences of ii in AA: This means that the MEX of both PP and QQ is ii.
There is 11 occurrence of ii in AA: We can place this ii either in PP or in QQ. If we place it in PP, the MEX of QQ becomes ii. However, the MEX of PP would atleast be (i+1)(i+1). This means that the MEX of PP and QQ can never be equal.
Implementation: We store the count of each number in array AA. Starting from 00, find the smallest number ii having count less than 22. All numbers less than ii can be distributed among PP and QQ such that there is atleast one occurrence in both. If ii has count 00, the answer is YES. If it has count 11, the answer is NO.